requirements-minimal.txt
# Minimal dependencies for Railway deployment
# This is a lightweight version that excludes optional database packages

# Core async framework
asyncio-contextmanager==1.0.0

# HTTP client
aiohttp==3.9.1

# ZeroMQ messaging
pyzmq==25.1.0

# Financial data
yfinance==0.2.32
requests==2.31.0

# Hashing
mmh3==4.0.1

# Serialization
msgpack==1.0.7

# Configuration
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0

# Logging
structlog==23.3.0
